✨ [feat] 게시글 삭제 기능 구현

+++ src/main/java/com/winner_cat/domain/article/controller/ArticleController.java
@@ -39,7 +39,7 @@ public class ArticleController {
     // 게시글 삭제
     @DeleteMapping("/{articleId}")
     public ResponseEntity<ApiResponse<?>> deleteArticle(
-            @PathVariable Long articleId) {
+            @PathVariable Integer articleId) {
         ResponseEntity<ApiResponse<?>> result = articleService.deleteArticle(articleId);
         return result;
     }
